This modifies the generation-service so that it will now generate SBs in the background using the same code as is used in the existing OT.
The mechanism is still very crude and needs further work.

It means that the generation-service is now aware of the location of the
files stored by the project-service. Note the docker-compose.yml
now needs updated in two places to point to the project store.
This should be refactored.

Some minor refactoring of the Project List (probably uneccessary)